+++ app/controllers/questions_controller.rb
@@ -1,38 +1,49 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
 
   def index
     @questions = Question.all
   end
 
   def show
+    @answer ||= question.answers.new
   end
 
-  def new
-  end
+  def new; end
 
   def edit
+    render :show, notice: "You can't edit someone else's question" unless current_user.author?(question)
   end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.new(question_params)
 
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your question successfully created.'
     else
       render :new
     end
   end
 
   def update
-    if question.update(question_params)
-      redirect_to @question
+    if current_user.author?(question)
+      if question.update(question_params)
+        redirect_to question
+      else
+        render :edit
+      end
     else
-      render :edit
+      render :show
     end
   end
 
   def destroy
-    question.destroy
+    if current_user.author?(question)
+      question.destroy
+      flash[:notice] = 'Your question was successfully deleted.'
+    else
+      flash[:notice] = "You cant't delete someone else's question"
+    end
     redirect_to questions_path
   end
